add api interfaces
