+++ src/config/Constants.ts
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export const BASE_URL = 'https://deezmos.azurewebsites.net/';
