+++ src/interface/schema/GraphDetailResponseInterface.ts
@@ -1,0 +1,8 @@
+import type GraphInterface from './GraphObjectInterface';
+
+interface GraphDetailResponse {
+	status: number;
+	message: string;
+	data: GraphInterface;
+}
+export default GraphDetailResponse;
