+++ src/interface/schema/GraphErrorResponseInterface.ts
@@ -1,0 +1,9 @@
+import type GraphInterface from './GraphObjectInterface';
+
+interface GraphValidationErrorResponse {
+	status: number;
+	message: string;
+	data: GraphInterface;
+}
+
+export default GraphValidationErrorResponse;
