+++ src/interface/schema/GraphObjectInterface.ts
@@ -1,0 +1,10 @@
+interface GraphInterface {
+	id: number;
+	name: string;
+	preview: string;
+	owner: number;
+	created: string;
+	updated: string;
+}
+
+export default GraphInterface;
